+++ download.py
@@ -1,135 +1,32 @@
 import os
-import re
-import shutil
-import subprocess
-from urllib.parse import urljoin
+import sys
 
-import requests
 import yaml
-from tqdm import tqdm
 
-from network import safe_get
-from api import (get_subscription_list, parse_subscription_list, get_user_info_by_code, get_timeline)
+from api import (
+    get_subscription_list,
+    parse_subscription_list,
+    get_user_info_by_code,
+    get_timeline,
+)
 
-from config import HEADERS, cfg, check_requirements, load_config
+from config import cfg, check_requirements, load_config
+
+try:
+    from downloader import sanitize_filename, download_and_merge
+except RuntimeError as e:
+    print(f"错误：{e}")
+    sys.exit(1)
 
 
 try:
     load_config()
 except FileNotFoundError:
     print("错误：未找到 config.yaml")
-    exit(1)
+    sys.exit(1)
 except yaml.YAMLError as e:
     print(f"配置文件格式错误: {e}")
-    exit(1)
-
-import sys
-
-
-# 入口前检查依赖
-ffmpeg_path = shutil.which("ffmpeg")
-if ffmpeg_path is None:
-    print("错误：未找到 ffmpeg。请先安装 ffmpeg 并确保其在系统 PATH 中。")
-    exit(1)
-
-
-# 清理文件名中的非法字符
-def sanitize_filename(filename):
-    return re.sub(r'[<>:"/\\|?*]', '_', filename).strip()
-
-
-# 创建目录
-def ensure_dir(path):
-    os.makedirs(path, exist_ok=True)
-
-
-# 下载并合并函数
-def download_and_merge(file_url, target_dir, output_name, url_type='m3u8'):
-    ensure_dir(target_dir)
-
-    # 如果是mp4文件，直接下载
-    if url_type == 'mp4':
-        output_path = os.path.join(target_dir, output_name)
-        print(f"[下载 MP4] {output_path}")
-        resp = safe_get(file_url, headers=HEADERS, stream=True)
-        resp.raise_for_status()
-        total_size = int(resp.headers.get("content-length", 0))
-        with open(output_path, "wb") as f, tqdm(
-                total=total_size, unit='B', unit_scale=True, desc=output_name
-        ) as pbar:
-            for chunk in resp.iter_content(1024 * 1024):
-                if chunk:
-                    f.write(chunk)
-                    pbar.update(len(chunk))
-        print(f"[下载完成] {output_path}")
-        return None
-
-    # 下载 m3u8 文本
-    r = safe_get(file_url, headers=HEADERS)
-    r.raise_for_status()
-    m3u8_text = r.text
-    m3u8_filename = os.path.join(target_dir, os.path.basename(file_url))
-    with open(m3u8_filename, "w", encoding="utf-8") as f:
-        f.write(m3u8_text)
-
-    # 判断 Master Playlist
-    lines = [l.strip() for l in m3u8_text.splitlines() if l.strip()]
-    if any(l.startswith("#EXT-X-STREAM-INF") for l in lines):
-        for i, l in enumerate(lines):
-            if l.startswith("#EXT-X-STREAM-INF"):
-                sub_1 = lines[i + 1]
-                base = file_url.rsplit("/", 1)[0] + "/"
-                sub_url = sub_1 if sub_1.startswith("http") else urljoin(base, sub_1)
-                return download_and_merge(sub_url, target_dir, output_name)
-
-    # 解析 TS 列表
-    base = file_url.rsplit("/", 1)[0] + "/"
-    ts_urls = [(l if l.startswith("http") else urljoin(base, l)) for l in lines if not l.startswith("#")]
-
-    # 下载 TS 并生成合并列表
-    filelist_path = os.path.join(target_dir, "filelist.txt")
-    with open(filelist_path, "w", encoding="utf-8") as list_f:
-        with tqdm(total=len(ts_urls), unit='ts', desc="TS 下载") as pbar:
-            for idx, ts in enumerate(ts_urls):
-                ts_name = f"{idx:04d}.ts"
-                ts_path = os.path.join(target_dir, ts_name)
-                try:
-                    resp = safe_get(ts, headers=HEADERS, stream=True)
-                    resp.raise_for_status()
-                except requests.exceptions.SSLError as e:
-                    print(f"[重试中] TS {idx} SSL 错误: {e}")
-                    resp = safe_get(ts, headers=HEADERS, stream=True)
-                    resp.raise_for_status()
-                with open(ts_path, "wb") as ts_f:
-                    for chunk in resp.iter_content(1024 * 1024):
-                        if chunk:
-                            ts_f.write(chunk)
-                list_f.write(f"file '{ts_name}'\n")
-                pbar.update(1)
-
-    # 合并 TS
-    output_path = os.path.join(target_dir, output_name)
-    cmd = [ffmpeg_path, "-y", "-f", "concat", "-safe", "0", "-i", filelist_path,
-           "-c", "copy", "-ignore_unknown", "-fflags", "+genpts", output_path]
-    try:
-        subprocess.run(cmd, check=True)
-    except subprocess.CalledProcessError as e:
-        print(f"警告：FFmpeg 合并失败，尝试使用重新编码方式: {e}")
-        cmd = [ffmpeg_path, "-y", "-f", "concat", "-safe", "0", "-i", filelist_path,
-               "-c:v", "libx264", "-c:a", "aac", "-ignore_unknown", "-fflags", "+genpts", output_path]
-        try:
-            subprocess.run(cmd, check=True)
-        except subprocess.CalledProcessError as e:
-            print(f"错误：FFmpeg 重新编码也失败了: {e}")
-            raise
-    print(f"[合并完成] {output_path}")
-
-    # 清理中间文件
-    for filename in os.listdir(target_dir):
-        if filename.endswith(".ts") or filename.endswith(".m3u8") or filename == "filelist.txt":
-            os.remove(os.path.join(target_dir, filename))
-    print(f"[清理] 中间文件已删除")
-    return None
+    sys.exit(1)
 
 
 if __name__ == "__main__":
